slugz/requirements.txt
python-slugify==8.0.4
rich==13.7.1